# server/openapi.json
{
  "components": {
    "schemas": {
      "APIKeyResponse": {
        "description": "Management representation. `key_prefix` is the non-secret display prefix;\nthe full secret is never stored, so it never appears here.",
        "properties": {
          "allowed_models": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "credit_limit": {
            "format": "double",
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quota_limit": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "rate_limit": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "usage_cost": {
            "format": "double",
            "type": "number"
          },
          "usage_tokens": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "key_prefix",
          "name",
          "enabled",
          "rate_limit",
          "quota_limit",
          "usage_tokens",
          "credit_limit",
          "usage_cost",
          "created_at"
        ],
        "title": "APIKeyResponse",
        "type": "object"
      },
...
      "CreatedAPIKeyResponse": {
        "description": "Creation response: the management fields plus the one-time full secret.",
        "properties": {
          "allowed_models": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "credit_limit": {
            "format": "double",
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "quota_limit": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "rate_limit": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "usage_cost": {
            "format": "double",
            "type": "number"
          },
          "usage_tokens": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "key_prefix",
          "name",
          "enabled",
          "rate_limit",
          "quota_limit",
          "usage_tokens",
          "credit_limit",
          "usage_cost",
          "created_at",
          "key"
        ],
        "title": "CreatedAPIKeyResponse",
        "type": "object"
      },
...
      "KeyListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/APIKeyResponse"
            },
            "type": "array"
          },
          "object": {
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "KeyListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminSession": {
        "description": "The admin session cookie. It also satisfies API-key auth on the routes that accept it.",
        "in": "cookie",
        "name": "srouter_admin_session",
        "type": "apiKey"
      },
      "bearerAuth": {
        "description": "The same API key in the `Authorization` header.",
        "scheme": "bearer",
        "type": "http"
      },
      "srouterApiKey": {
        "description": "An SRouter API key. `Authorization: Bearer <key>` carries the same credential.",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Multi-provider OpenAI and Anthropic compatible LLM gateway, as served by the Rust build (`server/`). One listener on `PORT` carries the API, the dashboard SPA, and the provider callbacks; there is no secondary OAuth listener. Every response uses snake_case field names (owner ruling 2026-10-07), and errors use the `{error:{message,type,code?,param?}}` envelope. Cloudflare Tunnel routes and `/v1/settings/fallbacks/*` are excluded from this build by owner ruling; they do not appear below and answer `404`.",
    "title": "SRouter API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/keys": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResponse"
                }
              }
            },
            "description": "Every key's management view; the secret is never returned."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "List API keys"
      },
      "post": {
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAPIKeyResponse"
                }
              }
            },
            "description": "The created key plus its one-time secret."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Create an API key"
      }
    },
    "/v1/keys/{id}": {
      "delete": {
        "parameters": [
          {
            "description": "Key id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The key was deleted; an unknown id answers `404`."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Revoke an API key"
      },
      "patch": {
        "parameters": [
          {
            "description": "Key id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            },
            "description": "The updated key."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Update an API key"
      }
    },
    "/v1/keys/{id}/credit": {
      "post": {
        "parameters": [
          {
            "description": "Key id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            },
            "description": "The updated key."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Add credit to an API key"
      }
    },
...
  }
}